Moved configuration page to pages/configuration directory.
