+++ src/pages/configuration/Sites.svelte
@@ -2,7 +2,7 @@
 	import { onMount } from "svelte";
 	import { hexToCSSFilter } from "hex-to-css-filter";
 
-	import { navigate } from "@/stores/page";
+	import { navigate } from "@/pages/configuration/pageStore";
 	import { getSiteConfigurations } from "@/utils/storage";
 	import Button from "@/components/Button.svelte";
