+++ src/pages/configuration/index.ts
@@ -1,5 +1,5 @@
 import { mount } from "svelte";
-import Main from "@/pages/Main.svelte";
+import Main from "@/pages/configuration/Main.svelte";
 
 mount(Main, {
 	target: document.getElementById("better-containers")!,
